[FROM-QUECTEL] Revert "ath11k: do not restore ASPM in case of single MSI vector"
This reverts commit 915a081ff307d61d6551d6c16b542e03775353c4.

Quectel:
Enabling ASPM has some undetermined issue in one MSI vector case,
so it is disabled by default in upstream.
However in this specific project, we did a large amount of tests and confirm
it is OK to enable it. So we will not upstream it.

Signed-off-by: neil.shi <neil.shi@quectel.com>
Cherry-picked-from: 8a797047843f ("Subject: [PATCH 38/42] Revert "ath11k: do not restore ASPM in case of single MSI vector"")
Signed-off-by: Eugen Hristev <eugen.hristev@collabora.com>
Signed-off-by: Cristian Ciocaltea <cristian.ciocaltea@collabora.com>