[NOT-FOR-UPSTREAM] Revert "Revert "wifi: ath11k: support hibernation""
The hibernation support was added in
166a490f59ac("wifi: ath11k: support hibernation"). It was reverted later
on 2f833e8948d6("Revert "wifi: ath11k: support hibernation"") as it was
causing regression is some lenovo devices [1][2].

There was stuttering issue reported on steamdeck [3] and initially people
thought that this commit is causing it on Wifi 6. But even after the
revert the issue is there. Hence the stuttering is unrelated to this
commit.

Let's revert the revert as without it ath11k doesn't hibernate correctly.

This reverts commit 2f833e8948d6c88a3a257d4e426c9897b4907d5a.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2301921
[2] https://bugzilla.kernel.org/show_bug.cgi?id=219196
[3] https://github.com/ValveSoftware/SteamOS/issues/1445#issuecomment-2546357312

[Resolve merge conflicts]
Signed-off-by: Muhammad Usama Anjum <usama.anjum@collabora.com>