[NOT-FOR-UPSTREAM] platform/x86: Update cache on ACPI AC events
Modifies current behavior of lenovo-wmi-capdata01 and lenovo-wmi-other
to re-cache the lenovo-wmi-capdata01 list upon AC connect/disconnect.
The LENOVO_CAPABILITY_DATA_01 data blocks change value based on the
presence of AC input to limit total output to under what the max power
draw of the battery is capable of.

This will be fixed in v6 of the upstream patches. As v5 was already
merged into frog/6.11, apply the changes directly.

Reported-by: Xino Ni <nijs1@lenovo.com>
Signed-off-by: Derek J. Clark <derekjohn.clark@gmail.com>