tools: usbip: Detect virtual port disconnect in vhci_driver
During listing of imported USB devices, do not simply skip all entries
which have the sysfs status 'not connected' or 'not assigned'.  Instead,
make use of the port record files, if available, to detect disconnect
states, and report them accordingly, along with the previous remote
connectivity details.

This will also serve as a prerequisite to introduce support for
automatic or manual reconnection.

Signed-off-by: Cristian Ciocaltea <cristian.ciocaltea@collabora.com>