[DONT-UPSTREAM] drm/amd/display: really dumb vrr resume fix
during dm resume, if the device is connected to a vrr monitor via
display port, the display is black and unusable until a replug of the
display port is done. Otherwise the system is working as normal.

As a temporary fix, re-trigger a dc link detect at the end of resume.
This fixes a 100% failure rate to a 100% success rate.
How or why is unclear as yet.

Don't Upstream: Need to figure out exactly why this works. I tried
various comibinations of integrating the link detect in to the resume
path to no avial. When we get time, we sould be able to figure out a
more integrated upstreamable solution, but for now this works.

Signed-off-by: Robert Beckett <bob.beckett@collabora.com>