[FROM-ML] platform/x86: msi-wmi-platform: Add quirk system
MSI uses the WMI interface as a passthrough for writes to the EC
and uses a board name match and a quirk table from userspace on
Windows. Therefore, there is no auto-detection functionality and
we have to fallback to a quirk table.

Introduce it here, prior to starting to add features to it.

Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>
Link: https://lore.kernel.org/all/20250511204427.327558-4-lkml@antheas.dev/
Signed-off-by: Matthew Schwartz <matthew.schwartz@linux.dev>
Signed-off-by: Derek J. Clark <derekjohn.clark@gmail.com>