[FROM-ML] platform/x86: msi-wmi-platform: Add platform profile through shift mode
MSI's version of platform profile in Windows is called shift mode.
Introduce it here, and add a profile handler to it.

It has 5 modes: sport, comfort, green, eco, and user.
Confusingly, for the Claw, MSI only uses sport, green, and eco,
where they correspond to performance, balanced, and low-power.
Therefore, comfort is mapped to balanced-performance, and user to
custom.

Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>
Link: https://lore.kernel.org/all/20250511204427.327558-6-lkml@antheas.dev/
Signed-off-by: Matthew Schwartz <matthew.schwartz@linux.dev>
Signed-off-by: Derek J. Clark <derekjohn.clark@gmail.com>