[NOT-FOR-UPSTREAM] drm/amd: Fix NULL ptr deref on vcn4.0.5
A NULL pointer deref occurs during startup because vcn 4.0.5 doesn't
have instances set up yes on the first pass for power gating.

That is because commit ecc9ab4e924b7 ("drm/amdgpu/vcn4.0.5: split code along instances")
is an implicit dependency for the VCN 4.0.5 startup when doing the
read back at end of start(). This is a complicated backport that
won't come back cleanly without a lot of other changes to amdgpu.

Instead fix the NULL pointer deref by moving the code into the scope
of the for loop. This can be dropped on a future rebase to 6.15 or
higher.

Link: https://lore.kernel.org/stable/20250521165421.293820-1-dnaim@cachyos.org/
Link: https://lore.kernel.org/stable/8db9b7cb-03ff-4aca-aafa-bcab4d1b5d82@cachyos.org/
Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>