[FOR-UPSTREAM] hid/lenovo-legos-hid: Make attributes static
Update the lenovo-legos-hid-config driver to statically store attributes
instead of using last_cmd pattern. This will permit the fwupd plugin for
the touchpad to query this info without probing the config interface.
It was found that probing this through fwupd with the driver loaded can
cause the MCU to lock. This allows fwupd to get the data, then it can
attach to the touchpad interface and grab exclusive access to it without
crashing the config driver and halting the MCU. In addition, some read
only attributes have been converted to uevent properties and the sysfs
paths for them have been dropped.

While at it, fix three bugs. Apply mutex when doing remove on the config
interface, add hid_hw_close/stop for all other interfaces, and restore
OS Mode attribute on resume.

Co-developed-by: Mario Limonciello <mario.limonciello@amd.com>
Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>
Signed-off-by: Derek J. Clark <derekojohn.clark@gmail.com>