[FOR-UPSTREAM] [SQUASHME] Bluetooth: btrrl: Don't check for fw signature
The fw versioning is not monotonic or anything useful we can use to
compare if a given fw is newer than another. Instead of trying to make
sure we have the correct fw or later to implement wobt, instead try to
set the ignore mask, and politely warn on error, assuming older fw.

Fixes: 3ee04c50b9b0 ("[FOR-UPSTREAM] Bluetooth: btrtl: add suspend handling to ignore BT_DIS")

Signed-off-by: Robert Beckett <bob.beckett@collabora.com>