[NOT-FOR-UPSTREAM] hid-asus-ally: disable wakeup attribute on N-Key device
With the current ROG Ally patchset, the MCU will send spurious events
when the ROG Ally is in s2idle while plugged in, which can cause the
system to wake unexpectedly.

There's really no reason the N-Key needs wakeup enabled, so just
disable the capability for now. This can be dropped when the issue is
root caused and fixed within the standard ROG Ally patchset.

Closes: https://github.com/ValveSoftware/SteamOS/issues/2119
Signed-off-by: Matthew Schwartz <matthew.schwartz@linux.dev>