[FOR-UPSTREAM] drm/amd/display: Initialise backlight level values from hw
Internal backlight levels are initialised from ACPI, but the values
are sometimes out of sync with the levels in effect until there has
been a read from hardware (eg by reading from sysfs).

This means that the first drm_commit can cause the levels to be set
to a different value than the actual starting one, which results in
a sudden change in brightness.

This path shows the problem (when the values are out of sync):

   amdgpu_dm_atomic_commit_tail()
   -> amdgpu_dm_commit_streams()
   -> amdgpu_dm_backlight_set_level(..., dm->brightness[n])

This patch calls the backight ops get_brightness explicitly
at the end of backlight registration to make sure dm->brightness[n]
is in sync with the hardware levels.

Signed-off-by: Vivek Das Mohapatra <vivek@collabora.com>