amdgpu_dm: Initialise the internal backlight level values from hw
Internal backlight levels are initialised from ACPI, but the values
appear to be static or at least out of date until somethiong triggers
a read from hardware (eg by reading from sysfs).

This means that the first drm_commit can cause the levels to be set
to a different value than the _actual_ starting one, which results
in a significant sudden change in brightness.

This patch calls the backight ops get_brightness member explicitly
at the end of backlight registration to make sure the internal
brightness value is what's actually in effect.