input: hid-lenovo-go: Forward all hid reports
While troubleshooting lenovo/issues#282 I discovered this driver to be
the culprit. Due to how we're handling the hid-reports we disregard it
will block any input reports from propagating to the other endpoints.
It also blocks creation of evdev nodes through the HID core from how
we're handing other endpoints in probe. To resolve this, do the following:

- Ensure HID quirks are set for the various endpoints in all applicable
  modes.
- Use HID_CONNECT_DEFAULT instead of HID_CONNECT_HIDRAW.
- Pass through all hid reports we aren't directly handling.

This patch will squash with no conflicts for 2/8 from the initial series,
"[FOR-UPSTREAM] input: hid-lenovo-go: Add Lenovo Legion Go Series HID Driver"

Signed-off-by: Derek J. Clark <derekjohn.clark@gmail.com>